backend/pyproject.toml
[project]
name = "quicknotes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "dotenv-python>=0.0.1",
    "fastapi>=0.116.1",
    "llama-index>=0.12.52",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-vector-stores-chroma>=0.4.2",
    "llama-index-vector-stores-postgres>=0.5.4",
    "llama-parse>=0.6.51",
    "mistralai>=1.9.3",
    "openai>=1.97.1",
    "psycopg[binary]>=3.2.9",
    "psycopg-pool>=3.2.6",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "pymupdf4llm>=0.0.27",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.0.0",
    "tiktoken>=0.9.0",
    "uvicorn>=0.35.0",
    "resend>=2.12.0",
]
